Moved counter to component
