+++ site/pages/LandingPage/index.js
@@ -6,17 +6,8 @@ import {
     Typography
 } from '@material-ui/core';
 
+import Countdown from '../../components/Countdown';
 import useStyles from './styles';
-
-const parseTimeLeft = timeleft => {
-    var days = Math.floor(timeleft / (1000 * 60 * 60 * 24));
-    var hours = Math.floor((timeleft % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-    var minutes = Math.floor((timeleft % (1000 * 60 * 60)) / (1000 * 60));
-    var seconds = Math.floor((timeleft % (1000 * 60)) / 1000);
-
-
-    return `${days}:${hours}:${minutes}:${seconds}`;
-};
 
 const goalComponent = team => {
     if (!team) {
@@ -28,7 +19,6 @@ const goalComponent = team => {
 
 const LandingPage = props => {
     const dayOfPlay = new Date("Nov 11, 2020 11:00:00").getTime();
-    const [now, setNow] = useState(new Date().getTime());
     const [team, setTeam] = useState(null);
     const [fetchingTeam, setFetching] = useState(false);
 
@@ -44,12 +34,6 @@ const LandingPage = props => {
                 .catch(err => console.log(err))
                 .finally(() => setFetching(false));
         }
-
-        const timerInterval = setInterval(() => {
-            setNow(new Date().getTime());
-        }, 1000);
-
-        return () => clearInterval(timerInterval);
     }, [fetchingTeam]);
 
     return (
@@ -57,7 +41,7 @@ const LandingPage = props => {
             <Typography variant="h1">Extra Life Slalom Atlanta</Typography>
             <Typography variant="h2">{goalComponent(team)}</Typography>
             <Typography variant="h2">Day of Play: November 7th 11:00 AM ET</Typography>
-            <Typography variant="h3">Countdown: {parseTimeLeft(dayOfPlay - now)}</Typography>
+            <Countdown dayOfPlay={dayOfPlay} />
         </Container>
     );
 };
